% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetActivityData.R
\docType{package}
\name{NetActivityData}
\alias{NetActivityData}
\title{NetActivityData: data required for getting the gene set scores with NetActivity package}
\description{
NetActivityData contains the weights from pre-trained shallow sparsely-connected
autoencoders. This data is required for getting the gene set scores with
NetActivity package.
}
